Don't enable mie.mtie during boot
If mie.mtie is enabled during the boot process then a timer interrupt
will be taken immediately after the switch to S-mode. This causes a
problem because the M-mode trap handler uses the same stack page as the
boot code, and resets sp to the base of the page and proceeds to store
values there, which may overwrite existing locals on the boot stack.

We can just leave timer interrupts disabled until the kernel requests to
set a timer interrupt, by which point booting is complete.